// frontend/src/app/models/election.models.ts
export interface ComparativeValue {
  v: number;
  d: string;
}
// ...
export interface Region {
  id: string;
  name: string;
  total: number;
  voted: number;
  partyVotes: { [key: string]: number };
  topParties?: { partyId: string, name: string, total: number, percentBp: number }[];
  discardedVotes?: number;
  noVotes?: number;
  totalPaper?: number;
  totalMachine?: number;
  avgTurnoutBp?: number;
  partyPercentsBp?: { [key: string]: number };
  comparisons?: { [key: string]: ComparativeValue[] };
}